media: i2c: max9271: Rename max9271 library driver
Support for the MAX9271 GMSL serializer was provided in the form of a
library driver, with the RDACM20 and RDACM21 camera module drivers using
the functions exported by the library.

In preparation to introduce an i2c subdevice driver to support the
MAX9271 serializer, rename the library driver from max9271 to
max9271-lib.

Signed-off-by: Jacopo Mondi <jacopo+renesas@jmondi.org>
Reviewed-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>