arm64: dts: GMSL: Adapt to the use max9271 driver
Currently the whole RDACM20/21 camera module was handled by a single
driver and a single device node entry was required in DTS to describe
it.

With the introduction of the max9271 subdevice driver the camera module
is now described by two device nodes, one for the serializer and one for
the image sensor connected to it.

Signed-off-by: Jacopo Mondi <jacopo+renesas@jmondi.org>