clk: core: use 64-bit integers
There are quite a few chips that provide or consume frequencies
upwards of the limit imposed by long integers on 32-bit systems.

The whole clocking subsystem is using long integers for storing
frequency rates, which limits the usage of these chips on 32-bit
systems.

To address this, change the internal integer type to u64, wrap
the new u64 API to provide compatibility with the old API (both
for providers, and consumers) all while checking for 64-bit
consumers passing frequencies past the long limit to long
providers, and 64-bit providers returning frequencies past the
long limit to long consumers.

Signed-off-by: Cosmin Tanislav <demonsingur@gmail.com>