v4l2: mc: fix endpoint iteration
When creating links from a subdev to a sink, the current
logic tries to iterate over the endpoints of dev's fwnode.

This might not be correct when the subdev uses a different
fwnode compared to the dev's fwnode.

If, when registering, the subdev's fwnode is not set,
the code inside v4l2_async_register_subdev will set it to
the dev's fwnode.

To fix this, just use the subdev's fwnode.

Signed-off-by: Cosmin Tanislav <demonsingur@gmail.com>
Fixes: 0d3c81e82da9 ("media: v4l2-mc: add v4l2_create_fwnode_links helpers")