media: rcar-vin: Simplify remote source type detection
The video source connected to a VIN instance can be either be a CSI-2
receiver, a CS-ISP or a parallel device. Each one requiring slightly
different configuration, sometimes reusing the same registers with
different meaning depending on what video source is used. The video
source type can change at run-time using the media device API.

This was introduced with R-Car Gen3 with two possible video sources and
have since been extended with one more. Instead of adding more flags
that needs to be set/cleared when changing links in the media graph add
functionality to use the media device to determine the video source
type.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>