media: i2c: maxim-serdes: remove serializer pipe stream-id configurability
The default pipe stream-id is not proper on serializers that have
some missing pipes, like max96717 which has missing x and y pipes,
as the stream-id for pipe z is 0x2. Remove the configuration and just
default to the virtual index of the pipe, as opposed to the real one.
There's no conflicts in deserializer if the stream-id is the same
between links.
Serializers that have multiple pipes do not support streaming using
the same stream id over multiple pipes.
Deserializers can only accept streams from a single stream id over
a link, exception being the ones that have automatic stream id
configuration, in which case the stream id does not even matter.
Either way, hardocde pipe to stream id 0 in deserializer and we'll
add automatic detection of stream id.