dt-bindings: net: pcs: renesas,rzn1-miic: Add renesas,miic-phylink-active-low property
Add a boolean DT binding property `renesas,miic-phylink-active-low` to the
RZN1 MIIC binding schema.

When present, this boolean indicates that the PHY-link signal provided by
the Ethernet switch is active-low. If omitted, the signal is assumed to
be active-high. This allows board DTs to express PHY-link polarity for SoC
variants that expose switch polarity control (for example, some RZ/T2H
variants that provide SWLINK[n]/SWLNK bits), while remaining compatible
with platforms that do not require an explicit property.

Signed-off-by: Lad Prabhakar <prabhakar.mahadev-lad.rj@bp.renesas.com>
Reviewed-by: Fabrizio Castro <fabrizio.castro.jz@renesas.com>