net: pcs: rzn1-miic: Add helper to adjust switch port speed and duplex
Add a new helper function miic_switchcore_adjust() which configures the
speed and duplex settings of MIIC switch ports. The function updates the
SWCTRL register to set the port speed (10/100/1000 Mbps) and the SWDUPC
register to configure duplex mode (half or full).

Only ports 0-2 are handled since these correspond to switch ports on the
hardware. This function will be used by switch drivers to ensure that
the MIIC configuration matches the link state negotiated by the MAC or
PHY.

Signed-off-by: Lad Prabhakar <prabhakar.mahadev-lad.rj@bp.renesas.com>
Reviewed-by: Fabrizio Castro <fabrizio.castro.jz@renesas.com>