dmaengine: sh: rz_dmac: make error interrupt optional
The Renesas RZ/T2H (R9A09G077) and RZ/N2H (R9A09G087) SoCs do not have
an error interrupt for the DMACs, and the current driver implementation
does not make much use of it.

To prepare for adding support for these SoCs, do not error out if the
error interrupt is missing.

Change-Id: I2ee307928133e59860c468118974012edc1f63b2
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>