clk: renesas: cpg-mssr: use read_status_reg()/write_control_reg() in PM
Re-use the status reading and control writing functions in the PM
.suspend_noirq() and .resume_noirq() callbacks.

For chips with status registers, this introduces a functional change
causing the .suspend_noirq() callback to use the values inside the
status registers to save the enabled clocks when suspending, matching
the .is_enabled() logic of the clocks.

This also fixes an issue when entering s2idle on RZ/T2H and RZ/N2H, as
the register layout in use by them was not accounted for, causing
aborts, and ensures that all layouts will be accounted for in the
future.

Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>