dt-bindings: timer: renesas,rz-mtu3: remove TCIU8 for RZ/G2{L,LC}
Based on the Page 528, Table 8.2 Interrupt mapping (7/13) in the User
Manual for the Renesas RZ/G2L and RZ/G2LC SoCs, channel 8 of the MTU3a
block does not have a TCIU8 interrupt, only a TCIV8 interrupt, as the
row where TCIU8 should have been is marked as reserved, and the GIC SPI
numbers stop at 212.

Remove the TCIU8 interrupt by setting maxItems to 43 for interrupts and
interrupt-names properties.

Change-Id: Ica4784852e15c0335d5d2826523966294b14b3b5
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>