arm64: dts: renesas: r9a07g044: remove TCIU8 interrupt from MTU3
Based on the Page 528, Table 8.2 Interrupt mapping (7/13) in the User
Manual for the Renesas RZ/G2L and RZ/G2LC SoCs, channel 8 of the MTU3a
block does not have a TCIU8 interrupt, only a TCIV8 interrupt, as the
row where TCIU8 should have been is marked as reserved, and the GIC SPI
numbers stop at 212.

Remove the TCIU8 interrupt. This does not cause any breakage as the
driver does not make use of the interrupts.

Change-Id: I7048d7e98673bc6d8582d603d7ed29dfe9d2a817
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>