pwm: rz-mtu3: simplify hwpwm to HW channel conversion
The MTU3 block has multiple HW channels, each with either 1 or 2 IOs
which can output a PWM signal. Each IO is exposed as a separate PWM
channel.

Previously, we did a complicated multi-step indirection to find the HW
channel for each PWM channel, where we hardcoded the base PWM number
for each HW channel (even though the base PWM number is just the sum of
the number of IOs for all previous HW channels), and we ended up
iterating through all channels to find the correct channel data anyway.

Remove struct rz_mtu3_channel_io_map and all its usages.

Create an intermediate IO numbering scheme which maps a PWM channel to
an index in channel_data[] and to primary/secondary IOs.

Make this mapping the source of truth for the number of PWM channels.

Use the mapping to remove the loop inside rz_mtu3_get_channel().

Add helpers to use this mapping and adapt the code accordingly.

Change-Id: Iffcc89f47022564bf29ef7a2273250f7038fd61e
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>