pwm: rz-mtu3: simplify hwpwm to HW channel conversion
The MTU3 block has multiple HW channels, each with either 1 or 2 IOs
which can be used to output a PWM signal. Each IO is exposed as a
separate PWM channel.

Previously, we did a complicated multi-step indirection to find the HW
channel for each PWM channel, where we hardcoded the base PWM number
for each HW channel (even though the base PWM number is just the sum of
the number of IOs for all previous HW channels), and we ended up
iterating through all of channels to find the correct channel data
anyway.

To simplify the setup, do the following.

Remove the base PWM number, the struct containing it, and its instance
from the channel data and from the channel_map array used to populate
the channel data.

While at it, also rename the channel_map to include the driver prefix.

Inside rz_mtu3_get_channel(), while iterating over the HW channels to
find the HW channel matching the given PWM channel number, keep track of
the base PWM number.

Add an output parameter to rz_mtu3_get_channel() named primary to find
out whether a PWM channel is the first IO of the HW channel by checking
the calculated base PWM number.

Change-Id: Iffcc89f47022564bf29ef7a2273250f7038fd61e
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>