pwm: rz-mtu3: move arrays into channel data
The user_count, enable_count, prescale, and period_cycles arrays are
indexed the same way as channel_data[], and channel_data[] is already
available everywhere these arrays are used.

Move them into channel_data[].

Remove the pointer subtraction used to derive the array index, as it is
not necessary anymore.

Change-Id: I56f47e5bb3b95d530e854b927e2bdb286e11a5d8
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>