mfd: rz-mtu3: do not use struct rz_mtu3_channel's dev member
The counter driver can use HW channels 1 and 2, while the PWM driver can
use HW channels 0, 1, 2, 3, 4, 6, 7.

The dev member is assigned both by the counter driver and the PWM driver
for channels 1 and 2, to their own struct device instance, overwriting
the previous value.

Depending on the probe order of the sub-drivers, the parent driver might
read either the dev pointer assigned by the counter driver, or the one
assigned by the PWM driver. The sub-drivers race to assign their own
struct device pointer to the same struct rz_mtu3_channel's dev member.

Both of these dev pointers have the same parent device, which means that
the struct rz_mtu3_priv can be retrieved from either of them, but this
behavior is inconsistent.

In other to fix this inconsistency, introduce a rz_mtu3_ch_to_priv()
helper that takes in a pointer to a struct rz_mtu3_channel and returns a
pointer to the struct rz_mtu3_priv.

This is possible because the struct rz_mtu3_channel contains its index
(channel_number) in the struct rz_mtu3's channels array, allowing us to
retrieve the pointer of the first channel, and then its container.

By doing this, we remove the dependency on sub-driver assigned data from
the parent driver.

Change-Id: I12054c61d5585598f6a95e197bbf55f5c650a04d
Fixes: 654c293e1687 ("mfd: Add Renesas RZ/G2L MTU3a core driver")
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>