mfd: rz-mtu3: do not use struct rz_mtu3_channel's dev member
The counter driver can use HW channels 1 and 2, while the PWM driver can
use HW channels 0, 1, 2, 3, 4, 6, 7.

The dev member is assigned both by the counter driver and the PWM driver
for channels 1 and 2, to their own struct device instance, overwriting
the previous value.

The dev member of struct rz_mtu3_channel is used by the counter
sub-driver for runtime PM and by the MFD driver to access its private
data.

Fix this inconsistency by introducing a rz_mtu3_ch_to_priv() helper
that takes in a pointer to a struct rz_mtu3_channel and returns a
pointer to the struct rz_mtu3_priv it belongs to.

This is possible because the struct rz_mtu3_channel contains its index
(channel_number) in the struct rz_mtu3's channels array, allowing us to
retrieve the pointer of the first channel, and then its container.

By doing this, we remove the dependency on sub-driver assigned data from
the parent driver.

Change-Id: I12054c61d5585598f6a95e197bbf55f5c650a04d
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>