pwm: rz-mtu3: add support for more prescale values
The MTU3 IP supports /1, /2, /4, /8, /16, /32, /64, /256, /1024 prescale
values, but we previously only supported /1, /4, /16 and /64 (powers of
4), for simplicity.

This limits the maximum period and duty-cycle granularity to a small
range. For example, at a 250 MHz input clock the maximum period is only
about 16.7 ms, which is too short for an observable LED blink.

0xffff * 64 * 1000000000 / 250000000 = 0xFFFF00 = 16776960 = ~16.7 ms

Add support for the rest of the prescale values, /2, /8, /32, /256, and
/1024, extending the period limit and precision.

0xffff * 1024 * 1000000000 / 250000000 = 0xFFFF000 = ~268.4 ms

To achieve this, program the TPSC2 field of the TCR2 and make the
following changes.

Change the internal prescale representation from powers of 4 to powers
of 2, which simplifies some of the calculations.

Introduce arrays that map powers of 2 to the TPSC and TPSC2 field
values, and add the rz_mtu3_pwm_prescale_to_tpsc() and
rz_mtu3_pwm_tpsc_to_prescale() functions to use those arrays.
Prescales <= /64 share the same field values for all channels, while
/256 and /1024 differ per channel.

Change the type of prescaled_period_cycles from u32 to u64 inside
rz_mtu3_pwm_calculate_prescale() to avoid overflows, as period_cycles
may exceed 48 bits and shifting it right by 16 can still leave a value
wider than 32 bits.

Also, since the hardware does not support /128 and /512 prescale values,
map these to /256 and /1024 respectively.

Change-Id: I25db4deecf68d47d48e3dcd52ff5bb8a9f21d5ee
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>