pwm: rz-mtu3: add support for more prescale values
The MTU3 IP supports /1, /2, /4, /8, /16, /32, /64, /256, /1024 prescale
values, but we previously only supported /1, /4, /16 and /64 (powers of
4), for simplicity.

This limits the period and duty cycle values we can use to a pretty
small value, which would not cause an observable blink on an LED, for
example.

0xffff * 64 * 1000000000 / 250000000 = 0xFFFF00 = 16776960 = ~16.7ms,

Add support for the rest of the prescale values, /2, /8, /32, /256, and
/1024, extending the period limit and precision.

0xffff * 1024 * 1000000000 / 250000000 = 0xFFFF000 = ~268.4ms

To achieve this, we have to program the TPSC2 field of the TCR2, and to
do the following changes.

Change the prescale value from powers of 4 to powers of 2, which
simplifies some of the calculations.

Introduce arrays that map from a power of 2 to the TPSC and TPSC2 fields
values and rz_mtu3_pwm_prescale_to_tpsc() and
and rz_mtu3_pwm_tpsc_to_prescale() functions to use those arrays.

The TPSC and TPSC2 values for prescales <= /64 are the same across the
channels, but for /256 and /1024 they are different enough to justifty
separate arrays that map from the HW channel number to the TPSC and
TPSC2 values.

To calculate the HW channel number, introduce a rz_mtu3_channel_number()
helper. This logic was previously present inline in multiple places
across the driver, but was removed.

Inside rz_mtu3_pwm_calculate_prescale(), change the type of
prescaled_period_cycles from u32 to u64 to avoid overflows since
period_cycles could contain values occupying more than 48 bits, leaving
more than 32 bits even after being shifted to the right by 16 (the
counter size).

Also, since the hardware does not support /128 and /512 prescale values,
map these to /256 and /1024 respectively.

Inside rz_mtu3_pwm_calculate_pv_or_dc(), remove the min() wrapper since
prescale values larger than the supported ones are rejected by
rz_mtu3_pwm_prescale_to_tpsc().

Change-Id: I25db4deecf68d47d48e3dcd52ff5bb8a9f21d5ee
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>