pwm: rz-mtu3: impose period restrictions
The counter is shared by all IOs of a HW channel, and we cannot clear
it from multiple sources, as the TCR register for each HW channel can
only select one clearing source between TGRA, TGRB, TGRC, and TGRD, or
the counter being cleared in another channel when synchronous clearing
is enabled.

Because of this hardware limitation, both IOs of a HW channel must share
the same period.

To provide some flexibility, allow requesting different periods for each
IO, with the following restrictions.

If the requested period is smaller than the already programmed period of
the sibling IO, return -EBUSY.

Otherwise, if the requested period is larger than the already programmed
period of the sibling IO, adjust the requested period to match it, and
adjust the duty cycle to not exceed it.

Since only one period is being used, always use TGRA for resetting the
counter, and program TGRA for secondary IOs too.

Change-Id: I85a144a5860c82dd211aad208b82646e832d04ca
Cc: stable@vger.kernel.org
Fixes: 254d3a727421 ("pwm: Add Renesas RZ/G2L MTU3a PWM driver")
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>