pwm: rz-mtu3: fix prescale check when enabling 2nd channel
enable_count is only incremented after rz_mtu3_pwm_config() is called
for the current PWM channel, causing prescale to not be checked if one
PWM channel is enabled and we're enabling the second PWM channel of the
same HW channel.

To handle this edge case, if the user_count of the HW channel is larger
than 1 and the sibling PWM channel is enabled, check that the new
prescale is not larger than the sibling's prescale.

If the new prescale is smaller than the sibling's prescale, use the
sibling's prescale.

The user_count check is necessary to ensure that we are indeed dealing
with a HW channel that has two IOs.

Change-Id: I9771a94649bef46b533457bf9cd1c095e9e6757d
Cc: stable@vger.kernel.org
Fixes: 254d3a727421 ("pwm: Add Renesas RZ/G2L MTU3a PWM driver")
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>