pwm: rz-mtu3: fix prescale check when enabling second IO
If one IO of a HW channel is already enabled the prescale check is
skipped when enabling the second IO.

The prescale check inside rz_mtu3_pwm_config() is only done if
enable_count is greater than 1.

enable_count is only incremented after the call to rz_mtu3_pwm_config(),
so it is still 1 when enabling the second IO of a HW channel, causing
the driver to incorrectly assume that the prescale can be changed.

As prescale is shared between IOs, it must not be changed if it would
affect the output of another PWM channel.

To fix this, take pwm->state.enabled into account when checking
enable_count. enable_count tracks enabled IOs, while pwm->state.enabled
indicates whether the current PWM is already enabled. Compare them to
determine whether the other IO is enabled and avoid changing the
prescale in that case.

Change-Id: I9771a94649bef46b533457bf9cd1c095e9e6757d
Cc: stable@vger.kernel.org
Fixes: 254d3a727421 ("pwm: Add Renesas RZ/G2L MTU3a PWM driver")
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>