pwm: rz-mtu3: impose period restrictions on second IO of HW channel
The counter is shared by all IOs of a HW channel, and we cannot clear
it from multiple sources, as the TCR register for each HW channel can
only select one clearing source between TGRA, TGRB, TGRC, and TGRD.

Because of this hardware limitation, both IOs of a HW channel must share
the same period.

Since only one period is being used, always use TGRA for resetting the
counter, and program TGRA for secondary IOs too.

To provide a bit of flexibility, use the largest period when both IOs
are enabled.

Change-Id: I85a144a5860c82dd211aad208b82646e832d04ca
Signed-off-by: Cosmin Tanislav <cosmin-gabriel.tanislav.xa@renesas.com>